Consolidate custom hooks
